move DomainModelHelper back to Not.Extensions
