+++ src/Not/Not.Domain/Base/DomainObject.cs
@@ -7,6 +7,6 @@ public abstract record DomainObject
 {
     protected string Combine(params object?[] values)
     {
-        return DomainHelper.Combine(values);
+        return DomainModelHelper.Combine(values);
     }
 }
